+++ subscription/mongodb/native/blocking-competing-consumer-strategy/src/main/java/org/occurrent/subscription/mongodb/spring/blocking/NativeMongoLeaseCompetingConsumerStrategy.java
@@ -72,6 +72,12 @@ public class NativeMongoLeaseCompetingConsumerStrategy implements CompetingConsu
         support.unregisterCompetingConsumer(collection, subscriptionId, subscriberId);
     }
 
+    @Override
+    public void releaseCompetingConsumer(String subscriptionId, String subscriberId) {
+        support.releaseCompetingConsumer(collection, subscriptionId, subscriberId);
+
+    }
+
     /**
      * Check whether a particular subscriber has the lock (access) to read events for the given subscription.
      *